Remove the last places in the code that depend on register_globals, to enable us to upgrade the prod server to php5.4+

+++ vote_thread.php
@@ -13,6 +13,9 @@ if ( !User::logged_in())
 	gdisp_end();}//end guest display
 else //elseallowed to edit
 {
+	if (!isset($_REQUEST['messageid']) die("messageid must be specified"); 
+	$messageid = (int)$_REQUEST['messageid'];
+	$vote = isset($_REQUEST['vote'])?$_REQUEST['vote']:"";
         mysql_query("delete from boardvotes where userid = '$idcookie' and messageid = '$messageid'");
         if ($vote == "y") {
             mysql_query("insert into boardvotes (userid, threadid, messageid, vote) values ('$idcookie',(select threadid from subboard where messageid = '$messageid'), '$messageid', 1)");
